Update Module_12_Lab.py
Updated with some minor changes.  Primarily I moved the print(my_pet) from main().

This was a very basic assignment and I chose to keep it simple.  There is no input validation, and I chose to keep modules limited.  Normally I would have an Initiations module to prompt the user for input, then the assorted executions, and a final output, but it seemed extraneous so I chose to not do it.